빌드 세팅 / 기타 버그 수정
1. Editor 관련 스크립트 Editor 폴더에 넣음

2. 자잘한 버그 수정
