+++ WAGTAIL/Assets/01_Scripts/00_Player/PlayerControls.inputactions
@@ -71,61 +71,6 @@
             ],
             "bindings": [
                 {
-                    "name": "WASD",
-                    "id": "f75fba1e-e2a9-437f-9d99-cbdad2b13510",
-                    "path": "2DVector",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Move",
-                    "isComposite": true,
-                    "isPartOfComposite": false
-                },
-                {
-                    "name": "Up",
-                    "id": "73178786-60a1-4a95-ae7d-1b7cd80c8056",
-                    "path": "<Keyboard>/w",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Move",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                },
-                {
-                    "name": "Down",
-                    "id": "de05e45a-4a72-424a-bc03-7d94c19606a5",
-                    "path": "<Keyboard>/s",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Move",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                },
-                {
-                    "name": "Left",
-                    "id": "11373927-61d4-4af9-823e-337f1fa0c545",
-                    "path": "<Keyboard>/a",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Move",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                },
-                {
-                    "name": "Right",
-                    "id": "147a3454-0977-418b-a024-6fb0f6efaeaf",
-                    "path": "<Keyboard>/d",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Move",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                },
-                {
                     "name": "ARROW",
                     "id": "a391c71f-bdc9-4510-9aa1-5717ba326851",
                     "path": "2DVector",
