+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class/Character.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEditor;
 
 [System.Serializable]
 public class MonsterPattern
