+++ WAGTAIL/Assets/01_Scripts/04_Manager/CameraManager.cs
@@ -1,13 +1,9 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Data;
 using System.Linq;
 using Cinemachine;
-using Unity.VisualScripting;
-using UnityEditor.Timeline;
 using UnityEngine;
-using UnityEngine.Rendering;
 
 public enum CameraType
 {
