--- WAGTAIL/Assets/01_Scripts/Enviroment Script/GrassTools/GrassPainterWindow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 81fd2c263a56f9d41a0b046e5bca5976
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
