+++ WAGTAIL/Assets/02_Prefabs/02_Object/Inter_Stonebridge 1.prefab
@@ -149,17 +149,17 @@ PrefabInstance:
     - target: {fileID: 457472801165828458, guid: 22e90733b5596ca43aa82d8e21034f6e,
         type: 3}
       propertyPath: m_Layer
-      value: 6
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 457472801849260143, guid: 22e90733b5596ca43aa82d8e21034f6e,
         type: 3}
       propertyPath: m_Layer
-      value: 6
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 457472802191387452, guid: 22e90733b5596ca43aa82d8e21034f6e,
         type: 3}
       propertyPath: m_Layer
-      value: 6
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 5484167771064285360, guid: 22e90733b5596ca43aa82d8e21034f6e,
         type: 3}
@@ -279,7 +279,7 @@ PrefabInstance:
     - target: {fileID: 5484167771064285363, guid: 22e90733b5596ca43aa82d8e21034f6e,
         type: 3}
       propertyPath: m_Layer
-      value: 6
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 5484167771064285363, guid: 22e90733b5596ca43aa82d8e21034f6e,
         type: 3}
