+++ WAGTAIL/Assets/06_Animations/00_Player/Throw.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7651c577778fa74fa97133025c49e28
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
