+++ WAGTAIL/Assets/99_Dummy/Environment/TerrainData/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9781f0bbb7d7ffc42b2202668cbfd649
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
