+++ WAGTAIL/Assets/99_Dummy/Environment/TerrainData/Editor/ExportTerrain.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2aa21697f2ef47499e929227fccb3a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
