+++ WAGTAIL/Assets/Plugins/FMOD/Cache/Editor/FMODStudioCache.asset
@@ -15,7 +15,7 @@ MonoBehaviour:
   Path: ../../IpariFMOD/IpariAudio/Build/Desktop/BGM.bank
   Name: BGM
   StudioPath: bank:/BGM
-  lastModified: 638305876761731146
+  lastModified: 638290070635154762
   FileSizes:
   - Name: Desktop
     Value: 10839872
@@ -169,7 +169,7 @@ MonoBehaviour:
   Path: ../../IpariFMOD/IpariAudio/Build/Desktop/Master.bank
   Name: Master
   StudioPath: bank:/Master
-  lastModified: 638305876761800965
+  lastModified: 638290070635154762
   FileSizes:
   - Name: Desktop
     Value: 2276
@@ -468,7 +468,7 @@ MonoBehaviour:
   - {fileID: -5999180023837473950}
   StringsBanks:
   - {fileID: 7503394392174117104}
-  cacheTime: 638305876761860815
+  cacheTime: 638290070635189752
   cacheVersion: 131605
 --- !u!114 &1656859512897261894
 MonoBehaviour:
@@ -884,7 +884,7 @@ MonoBehaviour:
   Path: ../../IpariFMOD/IpariAudio/Build/Desktop/SFX.bank
   Name: SFX
   StudioPath: bank:/SFX
-  lastModified: 638305876761860815
+  lastModified: 638290070635189752
   FileSizes:
   - Name: Desktop
     Value: 1312608
@@ -904,7 +904,7 @@ MonoBehaviour:
   Path: ../../IpariFMOD/IpariAudio/Build/Desktop/Master.strings.bank
   Name: Master.strings
   StudioPath: bank:/Master.strings
-  lastModified: 638305876761810946
+  lastModified: 638290070635154762
   FileSizes:
   - Name: Desktop
     Value: 2626
