+++ WAGTAIL/ProjectSettings/EditorBuildSettings.asset
@@ -12,8 +12,8 @@ EditorBuildSettings:
     path: Assets/00_Scenes/UI/Title.unity
     guid: 7a75f2efb6263044dbe3693c299fd0e5
   - enabled: 1
-    path: Assets/00_Scenes/Chapter/Chapter01.unity
-    guid: ff8bad5756e2e1c41aaf9e6773b2cc08
+    path: Assets/00_Scenes/GrayBoxing/Chapter01_Heejin 3.unity
+    guid: 71c6f3b47e6ce5445af82551bb1bc34e
   - enabled: 1
     path: Assets/00_Scenes/Chapter/Chapter02.unity
     guid: 4584d6f24a5978c4e9723e00e8a410f3
